move compatibility module to `falcon.util` package
