+++ falcon/testing/client.py
@@ -21,13 +21,12 @@ WSGI callable, without having to stand up a WSGI server.
 import warnings
 import wsgiref.validate
 
-from falcon import compat
-from falcon.compat import http_cookies
 from falcon.constants import MEDIA_JSON
 from falcon.testing import helpers
 from falcon.testing.srmock import StartResponseMock
-from falcon.util import CaseInsensitiveDict, http_date_to_dt, to_query_str
+from falcon.util import compat, CaseInsensitiveDict, http_date_to_dt, to_query_str
 from falcon.util import json as util_json
+from falcon.util.compat import http_cookies
 
 
 warnings.filterwarnings(
