+++ falcon/util/compat.py
