+++ tests/test_headers.py
@@ -4,8 +4,8 @@ from datetime import datetime
 import pytest
 
 import falcon
-from falcon import compat
 from falcon import testing
+from falcon.util import compat
 
 
 SAMPLE_BODY = testing.rand_string(0, 128 * 1024)
